requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
spacy>=3.7.0
nltk>=3.8.1
pdfplumber>=0.10.3
python-docx>=1.1.0
Pillow>=10.1.0
pytesseract>=0.3.10
rapidocr>=3.8.1
onnxruntime>=1.24.4
wordfreq>=3.1.1
PyMuPDF>=1.24.0
anthropic>=0.40.0
openpyxl>=3.1.2
aiofiles>=23.2.1
python-dotenv>=1.0.0
httpx>=0.26.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<5
